Making Loader non-generic
As the requirements of this project need no second call for any other resource we can make the Loader non-generic. Loader now only deals with the FeedModelContainer model which is the root object for feeds. Updating the test code to reflect this behavior changes. Also initiated some mapper objects to support encoding and decoding necessary for the test.